# docs/openapi/compute-service.yaml
...
components:
  schemas:
...
    Problem:
      properties:
        code:
          description: Discrete business error class.
          enum:
          - validation_failed
          - not_found
          - conflict
          - precondition_failed
          - unauthorized
          - forbidden
          - service_unavailable
          - internal_error
          - quota_exceeded
          type: string
        detail:
          type: string
        details:
          additionalProperties: {}
          type: object
        instance:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
...
    ResourceunitCreateInput:
      properties:
        description:
          type: string
        limits:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
        name:
          maxLength: 40
          minLength: 3
          pattern: ^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])?$
          type: string
        nodeSelector:
          additionalProperties:
            type: string
          type: object
        requests:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
      required:
      - name
      - requests
      type: object
    ResourceunitList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ResourceunitView'
          type: array
        total:
          format: int64
          type: integer
      required:
      - items
      - total
      type: object
    ResourceunitUpdateInput:
      properties:
        description:
          nullable: true
          type: string
        limits:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          nullable: true
          type: object
        nodeSelector:
          additionalProperties:
            type: string
          nullable: true
          type: object
        requests:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          nullable: true
          type: object
      type: object
    ResourceunitView:
      properties:
        createdAt:
          format: date-time
          type: string
        description:
          type: string
        id:
          format: uuid
          type: string
        limits:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
        name:
          type: string
        nodeSelector:
          additionalProperties:
            type: string
          type: object
        poolId:
          format: uuid
          type: string
        requests:
          additionalProperties:
            description: Kubernetes resource.Quantity.
            type: string
          description: Map of resource name (cpu, memory, nvidia.com/gpu, …) to resource.Quantity.
          type: object
        updatedAt:
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - name
      - poolId
      - requests
      - updatedAt
      type: object
info:
  description: REST API for resource pools, resource units, jobs, and services. Partitioned
    by Kubernetes namespace. RFC7807 Problem responses on errors.
  title: AxisML Compute Service API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/resource-pools/{pool}/resource-units:
    get:
      operationId: listResourceUnits
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      - description: Page size (1–200, default 50).
        in: query
        name: limit
        schema:
          format: int32
          type: integer
      - description: Number of items to skip.
        in: query
        name: offset
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceunitList'
          description: Page.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: List resource units in a pool
      tags:
      - resource-units
    post:
      operationId: createResourceUnit
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceunitCreateInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceunitView'
          description: Created.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Create a resource unit
      tags:
      - resource-units
  /api/v1/resource-pools/{pool}/resource-units/{unit}:
    delete:
      operationId: deleteResourceUnit
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      - description: Resource unit name.
        in: path
        name: unit
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted (idempotent).
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Delete resource unit
      tags:
      - resource-units
    get:
      operationId: getResourceUnit
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      - description: Resource unit name.
        in: path
        name: unit
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceunitView'
          description: Resource unit.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Get resource unit
      tags:
      - resource-units
    patch:
      operationId: updateResourceUnit
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      - description: Resource unit name.
        in: path
        name: unit
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceunitUpdateInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceunitView'
          description: Updated.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Patch resource unit
      tags:
      - resource-units
...
servers:
- description: Same-origin
  url: /
...